refactor(dedicated.pcc): rename files to respect convention
